Refactor: page wrapper stripped down to smaller template component and fix scroll behaviour
